Last-Modified tests for document resources
